BASISU_KTX2: implement deferred driver loading
